Give the draft page the report's own empty state
The draft shipped a placeholder body — a logo and "What can I help with?" —
while an empty report shows the agent picker and that selection's starter
questions. Two screens for the same moment, and the draft was the one
missing the controls that get a conversation started.

The empty state moves to ReportEmptyState and both render it, so an empty
report and a report that doesn't exist yet are the same screen. Its agent
chips, search, starter questions and training-mode variant come across
verbatim, test ids included; the chip/starter state that only ever served
this block (recency sort, the six-chip budget, the batched starter fetch)
moves with it, which is most of the 242 lines the report view loses.

The selection still lives in the prompt box — the empty state renders what
it publishes and emits clicks back, so on a draft the chips retarget the
report about to be created. A starter click sends, which on a draft is what
creates it, via a submitPrompt the box now exposes for controls outside it.

Verified in the sandbox: the two screens render the same controls, chips and
starters for the same agents; on the draft a chip retargets the scope, the
training variant replaces the picker, and a starter click creates exactly one
report whose first prompt is that starter. The four entry-point stories still
pass unchanged.
